Update 5.c
404notfound  ezy 5

+++ C/Easy/5.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int a = 10;
+    float a = 10;
     printf("Value of a: %f\n", a);
     return 0;
 }
